+++ deployments/chef/test/integration/with_default_systemd_dotnet_instrumentation/test.rb
@@ -1,8 +1,6 @@
 libsplunk_path = '/usr/lib/splunk-instrumentation/libsplunk.so'
-java_tool_options = '-javaagent:/usr/lib/splunk-instrumentation/splunk-otel-javaagent.jar'
-node_options = '-r /usr/lib/splunk-instrumentation/splunk-otel-js/node_modules/@splunk/otel/instrument'
-resource_attributes = 'splunk.zc.method=splunk-otel-auto-instrumentation-\d+\.\d+\.\d+-systemd,deployment.environment=test'
-otlp_endpoint = 'http://0.0.0.0:4317'
+resource_attributes = 'splunk.zc.method=splunk-otel-auto-instrumentation-\d+\.\d+\.\d+-systemd'
+otlp_endpoint = 'http://127.0.0.1:4317'
 dotnet_home = '/usr/lib/splunk-instrumentation/splunk-otel-dotnet'
 
 describe package('splunk-otel-auto-instrumentation') do
@@ -10,7 +8,7 @@ describe package('splunk-otel-auto-instrumentation') do
 end
 
 describe npm('@splunk/otel', path: '/usr/lib/splunk-instrumentation/splunk-otel-js') do
-  it { should be_installed }
+  it { should_not be_installed }
 end
 
 describe file('/etc/ld.so.preload') do
@@ -34,8 +32,8 @@ describe file('/usr/lib/splunk-instrumentation/instrumentation.conf') do
 end
 
 describe file('/usr/lib/systemd/system.conf.d/00-splunk-otel-auto-instrumentation.conf') do
-  its('content') { should match /^DefaultEnvironment="JAVA_TOOL_OPTIONS=#{java_tool_options}"$/ }
-  its('content') { should match /^DefaultEnvironment="NODE_OPTIONS=#{node_options}"$/ }
+  its('content') { should_not match /.*JAVA_TOOL_OPTIONS.*/ }
+  its('content') { should_not match /.*NODE_OPTIONS.*/ }
   its('content') { should match /^DefaultEnvironment="CORECLR_ENABLE_PROFILING=1"$/ }
   its('content') { should match /^DefaultEnvironment="CORECLR_PROFILER=\{918728DD-259F-4A6A-AC2B-B85E1B658318\}"$/ }
   its('content') { should match %r{^DefaultEnvironment="CORECLR_PROFILER_PATH=#{dotnet_home}/linux-x64/OpenTelemetry.AutoInstrumentation.Native.so"$} }
@@ -45,10 +43,10 @@ describe file('/usr/lib/systemd/system.conf.d/00-splunk-otel-auto-instrumentatio
   its('content') { should match /^DefaultEnvironment="OTEL_DOTNET_AUTO_HOME=#{dotnet_home}"$/ }
   its('content') { should match /^DefaultEnvironment="OTEL_DOTNET_AUTO_PLUGINS=Splunk.OpenTelemetry.AutoInstrumentation.Plugin,Splunk.OpenTelemetry.AutoInstrumentation"$/ }
   its('content') { should match /^DefaultEnvironment="OTEL_RESOURCE_ATTRIBUTES=#{resource_attributes}"$/ }
-  its('content') { should match /^DefaultEnvironment="OTEL_SERVICE_NAME=test"$/ }
-  its('content') { should match /^DefaultEnvironment="SPLUNK_PROFILER_ENABLED=true"$/ }
-  its('content') { should match /^DefaultEnvironment="SPLUNK_PROFILER_MEMORY_ENABLED=true"$/ }
-  its('content') { should match /^DefaultEnvironment="SPLUNK_METRICS_ENABLED=true"$/ }
+  its('content') { should_not match /.*OTEL_SERVICE_NAME.*/ }
+  its('content') { should match /^DefaultEnvironment="SPLUNK_PROFILER_ENABLED=false"$/ }
+  its('content') { should match /^DefaultEnvironment="SPLUNK_PROFILER_MEMORY_ENABLED=false"$/ }
+  its('content') { should match /^DefaultEnvironment="SPLUNK_METRICS_ENABLED=false"$/ }
   its('content') { should match /^DefaultEnvironment="OTEL_EXPORTER_OTLP_ENDPOINT=#{otlp_endpoint}"$/ }
 end