+++ deployments/chef/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## chef-v0.14.0
+
+- Initial support for [Splunk OpenTelemetry for .NET](https://github.com/signalfx/splunk-otel-dotnet) Auto
+  Instrumentation on Linux (x86_64/amd64 only):
+  - The .NET SDK is activated by default if the `with_auto_instrumentation` option is set to `true` and
+    `auto_instrumentation_version` is `latest` or >= `0.99.0`.
+  - To skip .NET auto instrumentation, configure the `with_auto_instrumentation_sdks` option without `dotnet`.
+
 ## chef-v0.13.0
 
 - Only copy the collector configuration file to `ProgramData` if the source exists.